fix: sqs message to update:score
